add overlay component
