+++ package.json
@@ -19,6 +19,7 @@
     "@codemirror/legacy-modes": "^6.0.1",
     "@codemirror/theme-one-dark": "^6.0.1",
     "@headlessui/react": "^1.7.4",
+    "@tanstack/react-query": "^4.16.1",
     "@theprelude/sdk": "^0.9.11",
     "classnames": "^2.3.2",
     "codemirror": "^6.0.1",
