+++ src/components/overlays/servers/connected-server.tsx
@@ -1,6 +1,6 @@
 import classNames from "classnames";
-import useAuthStore, { selectIsConnected } from "../../hooks/auth-store";
-import useNavigationStore from "../../hooks/navigation-store";
+import useAuthStore, { selectIsConnected } from "../../../hooks/auth-store";
+import useNavigationStore from "../../../hooks/navigation-store";
 import styles from "./servers.module.css";
 
 const ConnectedServer = () => {
