+++ src/components/overlays/servers/servers.module.css
@@ -1,64 +1,3 @@
-/* Blacked overlay */
-
-.overlay {
-  position: absolute;
-  height: 100vh;
-  width: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
-}
-
-/* Server Panel */
-
-.servers {
-  position: absolute;
-  display: flex;
-  flex-direction: column;
-  right: 0;
-  padding: 1.5rem;
-  width: 25%;
-  height: 100%;
-  background-color: var(--color-secondary-30);
-  border: 1px solid var(--color-secondary-10);
-}
-
-.closeIcon {
-  height: 1.25rem;
-  width: 1.25rem;
-  margin-left: auto;
-}
-
-.closeIcon:hover .icon {
-  fill: #fff;
-  cursor: pointer;
-}
-
-.content {
-  display: flex;
-  flex-direction: column;
-  row-gap: 1.5rem;
-}
-
-.titleBar {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  column-gap: 0.75rem;
-}
-
-.icon {
-  fill: var(--color-primary-10);
-}
-
-.title {
-  margin: 0;
-}
-
-.description {
-  margin: 0;
-  font: var(--text-base);
-  line-height: 1.5rem;
-}
-
 .selection {
   display: flex;
   flex-direction: row;
@@ -155,7 +94,7 @@ want to handle storing server input data*/
 
 .disconnect:hover {
   background-color: var(--color-darkAlert);
-  border: 1px solid var(--color-darkAlert); 
+  border: 1px solid var(--color-darkAlert);
 }
 
 /* Connected Server Overlay */
