+++ src/components/overlays/tests/test-catalog.module.css
@@ -2,40 +2,6 @@ button:focus {
   outline: none;
 }
 
-.overlay {
-  position: absolute;
-  height: 100vh;
-  width: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
-  display: flex;
-  flex-direction: row;
-}
-
-.backdrop {
-  flex: 1;
-}
-
-.panel {
-  z-index: 100;
-  position: absolute;
-  display: flex;
-  flex-direction: column;
-  padding: 0.75rem 0;
-  width: 475px;
-  height: 100%;
-  background-color: var(--color-secondary-40);
-  border-left: 1px solid var(--color-secondary-10);
-  overflow: auto;
-}
-
-.panel.left {
-  left: 0;
-}
-
-.panel.right {
-  right: 0;
-}
-
 .test {
   margin-bottom: 0.25rem;
 }
@@ -67,61 +33,6 @@ button:focus {
   transform: rotate(0deg);
 }
 
-.close {
-  background: none;
-  border: none;
-  margin-left: auto;
-  padding: 0 0.5rem 0 0;
-}
-
-.close > svg {
-  width: 1rem;
-  height: 1rem;
-  fill: var(--color-primary-10);
-}
-
-.close:hover {
-  cursor: pointer;
-}
-
-.close > svg:hover {
-  fill: var(--white);
-}
-
-.title {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-}
-
-.legend {
-  font-size: var(--text-xxl);
-  padding: 0.5rem;
-  color: var(--white);
-  display: flex;
-  flex-direction: row;
-  gap: 1rem;
-  padding: 0 0.75rem;
-  font-weight: 300;
-}
-
-.loaderIcon {
-  width: 1rem;
-  height: 1rem;
-  fill: var(--white);
-  animation: loader 1.2s linear infinite;
-  margin: auto 0;
-}
-
-@keyframes loader {
-  0% {
-    transform: rotate(0deg);
-  }
-  100% {
-    transform: rotate(360deg);
-  }
-}
-
 .helpIcon {
   width: 1rem;
   height: 1rem;
@@ -132,12 +43,6 @@ button:focus {
 .helpIcon:hover {
   fill: var(--white);
   cursor: pointer;
-}
-
-.description {
-  font: var(--text-sm);
-  padding: 0 0.75rem;
-  margin: 0.5rem 0 1.125rem;
 }
 
 .variants {
